renamed Payments to Plans, fix pages index.jsx issue
