+++ src/components/Main.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import About from '../pages/About';
 import Home from '../pages/Home';
 import News from '../pages/News';
-import Payments from '../pages/Payments';
+import Plans from '../pages/Plans';
 import StockDetails from '../pages/StockDetails';
 import Watchlist from '../pages/Watchlist';
 
@@ -12,7 +12,7 @@ const Main = () => {
         <>
             <Route path='/' exact component={Home} /> {/*Raghav*/}
             <Route path='/watchlist' component={Watchlist} /> {/*Weiyoung*/}
-            <Route path='/plans' component={Payments} /> {/*Himanshu*/}
+            <Route path='/plans' component={Plans} /> {/*Himanshu*/}
             <Route path='/about' component={About} /> {/*Weiyoung*/}
             <Route path='/news' component={News} /> {/*Paul*/}
             {/*<Route path="/help" component={Help}/> /!*Himanshu*!/*/}
