+++ src/components/StockSearchBar.jsx
