+++ src/pages/Plans.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import PageTemplate from "../../components/pagetemplate/PageTemplate";
-import PaymentPlanList from "../../components/PaymentPlanList";
-import ReviewCardList from "../../components/ReviewCardList";
-import CustomHeading from "../../components/CustomHeading";
+import PageTemplate from "../components/pagetemplate/PageTemplate";
+import PaymentPlanList from "../components/PaymentPlanList";
+import ReviewCardList from "../components/ReviewCardList";
+import CustomHeading from "../components/CustomHeading";
 import {Divider, useColorModeValue} from "@chakra-ui/react";
 
-const Payments = () => (
+const Plans = () => (
     <PageTemplate bgColor={useColorModeValue("brand.300", "brand.800")}>
         <CustomHeading>Select A Plan</CustomHeading>
         <PaymentPlanList/>
@@ -15,4 +15,4 @@ const Payments = () => (
     </PageTemplate>
 )
 
-export default Payments;
+export default Plans;
