+++ src/pages/StockDetails.jsx
@@ -1,13 +1,13 @@
 import React, {useEffect} from "react";
-import PageTemplate from "../../components/pagetemplate/PageTemplate";
+import PageTemplate from "../components/pagetemplate/PageTemplate";
 import {useColorModeValue} from "@chakra-ui/react";
-import StockInformation from "../../components/StockInformation";
-import StockNews from "../../components/StockNews";
-import StockPredictionDetails from "../../components/StockPredictionDetails";
-import StockPredictionOverview from "../../components/StockPredictionOverview";
-import {usePrediction} from "../../contexts/usePredictions";
-import {useStockSymbol} from "../../contexts/useStockInfo";
-import {useStockNews} from "../../contexts/useStockNews";
+import StockInformation from "../components/StockInformation";
+import StockNews from "../components/StockNews";
+import StockPredictionDetails from "../components/StockPredictionDetails";
+import StockPredictionOverview from "../components/StockPredictionOverview";
+import {usePrediction} from "../contexts/usePredictions";
+import {useStockSymbol} from "../contexts/useStockInfo";
+import {useStockNews} from "../contexts/useStockNews";
 
 const StockDetails = ({match}) => {
     const tickerId = match.params.tickerId;
