+++ src/pages/Watchlist.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import PageTemplate from "../../components/pagetemplate/PageTemplate";
-import CustomHeading from "../../components/CustomHeading";
+import PageTemplate from "../components/pagetemplate/PageTemplate";
+import CustomHeading from "../components/CustomHeading";
 import {Box, Center, Text, useColorModeValue} from "@chakra-ui/react";
-import {useUser} from "../../contexts/useUser";
+import {useUser} from "../contexts/useUser";
 
 const CustomBox = ({children, ...otherProps}) => {
     const bgColor = useColorModeValue("brand.300", "brand.700")
